+++ src/main/java/org/example/tryonx/admin/ask/dto/AskListDto.java
@@ -1,16 +1,17 @@
-package org.example.tryonx.admin.dto;
+package org.example.tryonx.admin.ask.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.tryonx.enums.Size;
 
-import java.time.LocalDateTime;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AskAnswerRequestDto {
+public class AskListDto {
     private Long askId;
-    private String answer;
+    private String title;
+    private String productName;
+    private Size size;
+    private String imgUrl;
 }
